Add oneByOne adds numbers input 1 char at a time
The oneByOne function allows the user apply basic mathematical
operations to single digit numbers. The function asks the user for a
single value, stores it on the stack if it is a number.
Then if it is one of the four basic mathematical operators, and there
are at least 2 numbers in the stack the operator is applied.
The answer is pushed onto the stack. When the user types q to quit
the value on top of the stack is returned.

The oneByOne function allows the program to function as a single digit
calculator that the user can utilize to apply the four basic
mathematical operations to the numbers that are entered.