Refactor math operation switch to another function
The switch statement that selected the math operation to perform was
moved to a new function that takes in the stack and the char entered
by the user. The function computes the answer stores it in the stack
and returns the stack to the oneByOne function.

The selectOperation function modularizes the code and makes it easier to
read the oneByOne function.